// src/model/login.ts
/*
 * @Author: Miya
 * @Date: 2020-08-07 18:48:24
 * @LastEditors: Miya
 * @LastEditTime: 2020-08-07 18:48:38
 * @Description: 登录表单模型
 * @FilePath: \Single-Search\src\model\login.ts
 */
export default interface LoginInterface {
  username: string;
  password: string;
  remember: boolean;
}